♻ refactor: ellipsis 컴포넌트 사용, 모달 props 명 수정
