+++ grass-diary/src/pages/Diary/modal/CompleteDeleteModal.jsx
@@ -38,8 +38,8 @@ const styles = stylex.create({
   },
 });
 
-const CompleteDeleteModal = ({ show }) => {
-  const closeModal = () => show(false);
+const CompleteDeleteModal = ({ setter }) => {
+  const closeModal = () => setter(false);
 
   return (
     <div {...stylex.props(styles.background)}>
